--- a/.env-example
+++ b/.env-example
@@ -11,7 +11,7 @@ AWS_S3_SECRET_ACCESS_KEY=
 ## DB 생성 시 루트 유저로 생성할 유저의 정보
 DB_USER=pg
 DB_PASSWORD=pg
-
+DB_MAX_CONNECTIONS=10
 ## 개발 단계에서는 아무 string이라도 상관 없음
 JWT_ACCESS_SECRET=acsecret
 JWT_REFRESH_SECRET=rfsecret
@@ -30,12 +30,12 @@ FRONT_END_URL=http://localhost:5173
 MAIL_FROM=  # example@naver.com 뒤에 메일 주소까지 전체 다 작성
 AWS_S3_REGION=ap-northeast-2
 AWS_S3_BUCKET_NAME=
-SERIES_ID=1 # 문자중계 크롤링 관련 | 1 => 정규시즌 | 3 => 준플레이오프 | 4 => 와일드카드 결정전 | 5 => 플레이오프 | 7 => 한국시리즈
+SERIES_ID=0 # 문자중계 크롤링 관련 |0, 1 => 정규시즌 | 3 => 준플레이오프 | 4 => 와일드카드 결정전 | 5 => 플레이오프 | 7 => 한국시리즈
 
-SWAGGER_USER=#배포시에만 사용
-SWAGGER_PW=#배포시에만 사용
+SWAGGER_USER=#배포 환경에서만 사용
+SWAGGER_PW=#배포 환경에서만 사용
 
-SLACK_WEBHOOK_URL=#프러덕션 환경에서만 필요
+SLACK_WEBHOOK_URL=#배포 환경에서만 필요
 GOOGLE_CLIENT_ID=
 GOOGLE_CLIENT_SECRET=
 KAKAO_CLIENT_ID=