--- a/src/common/guard/access-token.guard.ts
+++ b/src/common/guard/access-token.guard.ts
@@ -23,13 +23,14 @@ export class AccessTokenGuard implements CanActivate {
       throw new UnauthorizedException('토큰 형식 오류');
     }
 
-    const { user, payload } = await this.jwtStrategy.validateTokenAndGetUser(
-      token,
-      'access',
-    );
+    const payload = await this.jwtStrategy.validateToken(token, 'access');
+    const user = await this.jwtStrategy.checkUser(payload.id);
 
-    req.user = user;
-    req.token = token;
+    if (!user) {
+      throw new UnauthorizedException('유효하지 않은 사용자입니다.');
+    }
+
+    req.user = { ...user, email: payload.email };
     req.tokenType = payload.type;
 
     if (req.tokenType !== 'ac') {