--- a/src/core/config/database.config.ts
+++ b/src/core/config/database.config.ts
@@ -13,5 +13,8 @@ export const getDatabaseConfig = (
     password: configService.get<string>('DB_PASSWORD'),
     autoLoadEntities: true,
     synchronize: false,
+    extra: {
+      max: 25,
+    },
   };
 };