--- a/src/modules/auth/auth.controller.ts
+++ b/src/modules/auth/auth.controller.ts
@@ -300,10 +300,10 @@ export class AuthController {
   async deleteSocialLink(
     @Param('provider', ProviderParamCheckPipe)
     provider: SocialProvider,
-    @UserDeco() user: User,
+    @UserDeco('id') userId: number,
   ) {
     await this.accountService.unlinkSocial({
-      userId: user.id,
+      userId,
       provider,
     });
   }
@@ -336,8 +336,8 @@ export class AuthController {
       },
     },
   })
-  async checkRefreshToken(@UserDeco() user: User) {
-    return await this.accountService.checkUserAgreedRequiredTerm(user.id);
+  async checkRefreshToken(@UserDeco('id') userId: number) {
+    return await this.accountService.checkUserAgreedRequiredTerm(userId);
   }
 
   /** 엑세스 토큰 재발급 */
@@ -349,8 +349,14 @@ export class AuthController {
     type: AccessTokenResDto,
     description: '새로운 엑세스 토큰 발급',
   })
-  async reissueAcToken(@UserDeco() user: User): Promise<AccessTokenResDto> {
-    const { email, id, support_team } = user;
+  async reissueAcToken(
+    @UserDeco('id') userId: number,
+  ): Promise<AccessTokenResDto> {
+    const { email, id, support_team } = await this.authService.getUserForAuth(
+      userId,
+      { support_team: { id: true, name: true } },
+      { support_team: true },
+    );
     const acToken = this.authService.issueToken({ email, id }, 'access');
     return { acToken, teamId: support_team.id, teamName: support_team.name };
   }