## Description
### 1. Access/Refresh Guard 캐시 전략 개선
- 기존: 인증마다 DB 직접 조회
- 변경: Redis로 먼저 조회 → 없으면 DB 조회 후 Redis에 캐싱
- 효과: DB 부하 감소, 평균 응답 시간 단축

### 2. game.date 컬럼 인덱싱
- `/games/daily` API에서 date 필터 자주 사용
- 변경: date 컬럼에 인덱스 추가
- 효과: 날짜 범위 조회 속도 개선

### 3. `/users/me` API 병렬 처리 도입
- 기존: 순차 처리
- 변경: `Promise.all`로 병렬 처리

### 4. DB 커넥션 풀 사이즈 조정
- 목표: 병렬 처리 시 커넥션 부족/과부하 병목 완화
- 로컬: 45개로 실험해 큰 성능 향상 확인
- 운영: 리소스 제약 있어 15 전후로 타협 (무리하게 높이면 역효과)

### 5. Access-token 검증만 수행 (DB/Redis 접근 X)
- 기존: 토큰 확인 시 Redis or DB에서 추가 조회
- 변경: 단순 토큰 시그니처만 검증 (JWT verify) → DB/캐시 접근 제거
- 효과: 관련 API에서 DB/Redis 부하 극감, 응답 속도 급상승

## 성능 비교 표 (로컬환경)
- 로컬 환경에서 최대한 부하를 주는 방식으로 테스트

| 구분              | 개선 전       | 개선 후 v1 (1~4) | 개선 후 v2 (1~5) |
|-------------------|--------------|------------------|------------------|
| 평균 응답 시간    | 4,500 ms     | 193 ms           | 113 ms           |
| P95 응답 시간     | 8,000 ms     | 1,600 ms         | 1300 ms           |
| 최대 처리량 (TPS) | 765          | 1,685            | 1,948            |
| 가상 유저 완료율  | 3.2%        | 25.5%           | 31.3%           |


## 요약
- **단계별로 누적된 개선**을 통해 평균 응답시간/성공률/처리량 모두 극적으로 향상  
- 로컬 vs 운영환경이 달라도, “캐싱 + 병렬처리 + 커넥션 최적화 + 토큰검증 단순화” 등  
  **핵심 로직**은 그대로 적용 가능 
- 추후 더 높은 트래픽이 들어오고 DB 커넥션에서 병목이 발생한다면 RDS 인스턴스 사양 업그레이드 및 최대 커넥션을 증가시켜 대응가능